query optimization for user
