+++ internal/gophermart/db/postgres.go
@@ -1,12 +1,8 @@
 package db
 
 import (
-	"context"
 	"database/sql"
 	"log/slog"
-	"musthave_tpl/internal/gophermart"
-	"musthave_tpl/internal/models"
-	"time"
 
 	sq "github.com/Masterminds/squirrel"
 
@@ -55,240 +51,4 @@ func (ps *PostgresStorage) init() error {
 	`
 	_, err := ps.db.Exec(query)
 	return err
-}
-
-func (ps *PostgresStorage) UserExist(user models.User) (bool, error) {
-	row := ps.db.QueryRow(`SELECT COUNT(*) FROM users WHERE login = $1`, user.Login)
-	var count int64
-	err := row.Scan(&count)
-	if err != nil {
-		return false, err
-	}
-	if count != 0 {
-		return false, nil
-	}
-	return true, nil
-}
-
-func (ps *PostgresStorage) AddUser(user models.User) error {
-	exist, err := ps.UserExist(user)
-	if err != nil {
-		return err
-	} else if !exist {
-		return &gophermart.UsedLoginError{}
-	}
-	queryAddUser := ps.pSQL.Insert(
-		"users",
-	).Columns(
-		"login",
-		"password",
-	).Values(user.Login, user.Password)
-
-	_, err = queryAddUser.Exec()
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (ps *PostgresStorage) UserByLogin(login string) (*models.User, error) {
-	var (
-		userLogin     string
-		userPassword  string
-		userBalance   sql.NullFloat64
-		userWithdrawn sql.NullFloat64
-	)
-	query := ps.pSQL.Select(
-		"login", "password", "balance", "withdrawn",
-	).From(
-		"users",
-	).Where(sq.Eq{"login": login})
-
-	row := query.QueryRow()
-	if err := row.Scan(&userLogin, &userPassword, &userBalance, &userWithdrawn); err != nil {
-		return nil, err
-	}
-	return &models.User{
-		Login:     userLogin,
-		Password:  userPassword,
-		Balance:   userBalance.Float64,
-		Withdrawn: userWithdrawn.Float64,
-	}, nil
-}
-
-func (ps *PostgresStorage) OrdersByUser(login string) ([]models.Order, error) {
-	query := ps.pSQL.Select(
-		"id", "user_login", "status", "accrual", "uploaded_at",
-	).From(
-		"orders",
-	).Where(
-		sq.Eq{"user_login": login},
-	).OrderBy("uploaded_at ASC")
-
-	rows, err := query.Query()
-	if err != nil {
-		return nil, err
-	}
-	if rows.Err() != nil {
-		return nil, rows.Err()
-	}
-	var orders []models.Order
-	for rows.Next() {
-		var (
-			id         int64
-			userLogin  string
-			status     string
-			accrual    float64
-			uploadedAt time.Time
-		)
-		err := rows.Scan(&id, &userLogin, &status, &accrual, &uploadedAt)
-		if err != nil {
-			return nil, err
-		}
-		orders = append(orders, models.Order{
-			User:       models.User{Login: userLogin},
-			Number:     id,
-			Status:     models.OrderStatus(status),
-			Accrual:    accrual,
-			UploadedAt: uploadedAt,
-		})
-	}
-	return orders, nil
-}
-
-func (ps *PostgresStorage) AddOrder(order models.Order) error {
-	queryUser := ps.pSQL.Select(
-		"user_login",
-	).From(
-		"orders",
-	).Where(sq.Eq{"id": order.Number})
-
-	rows, err := queryUser.Query()
-	if err != nil {
-		return err
-	}
-	if rows.Err() != nil {
-		return err
-	}
-	if rows.Next() {
-		var login string
-		err := rows.Scan(&login)
-		if err != nil {
-			return err
-		}
-		if login == order.User.Login {
-			return &gophermart.OrderExists{}
-		}
-		return &gophermart.OrderUsed{}
-	}
-	ctx := context.Background()
-	tx, err := ps.db.BeginTx(ctx, nil)
-	if err != nil {
-		return err
-	}
-
-	orderQuery := ps.pSQL.Insert(
-		"orders",
-	).Columns(
-		"id",
-		"user_login",
-		"status",
-		"accrual",
-		"uploaded_at",
-	).Values(order.Number, order.User.Login, order.Status, order.Accrual, time.Now())
-
-	_, err = orderQuery.ExecContext(ctx)
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-
-	balanceQuery := ps.pSQL.Update(
-		"users",
-	).Set(
-		"balance",
-		sq.Expr("balance + ?", order.Accrual),
-	).Where(
-		sq.Eq{"login": order.User.Login},
-	)
-
-	_, err = balanceQuery.ExecContext(ctx)
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-	err = tx.Commit()
-	return err
-}
-
-func (ps *PostgresStorage) AddTransaction(transaction models.Transaction) (*models.Transaction, error) {
-	ctx := context.Background()
-	tx, err := ps.db.BeginTx(ctx, nil)
-	if err != nil {
-		return nil, err
-	}
-	queryWithdraw := ps.pSQL.Insert(
-		"withdrawals",
-	).Columns(
-		"user_login",
-		"order_id",
-		"amount",
-		"processed_at",
-	).Values(transaction.User, transaction.ID, transaction.Amount, transaction.Date)
-
-	_, err = queryWithdraw.ExecContext(ctx)
-	if err != nil {
-		tx.Rollback()
-		return nil, err
-	}
-	queryUser := ps.pSQL.Update(
-		"users",
-	).Set(
-		"balance",
-		sq.Expr("balance - ?", transaction.Amount),
-	).Set(
-		"withdrawn",
-		sq.Expr("withdrawn + ?", transaction.Amount),
-	).Where(sq.Eq{"login": transaction.User})
-	_, err = queryUser.ExecContext(ctx)
-	if err != nil {
-		tx.Rollback()
-		return nil, err
-	}
-	err = tx.Commit()
-	return nil, err
-}
-
-func (ps *PostgresStorage) TransactionsByUser(login string) ([]models.Transaction, error) {
-	query := ps.pSQL.Select(
-		"user_login",
-		"order_id",
-		"amount",
-		"processed_at",
-	).From(
-		"withdrawals",
-	).Where(sq.Eq{"user_login": login})
-
-	rows, err := query.Query()
-	if err != nil {
-		return nil, err
-	}
-	if rows.Err() != nil {
-		return nil, err
-	}
-	var transactions []models.Transaction
-	for rows.Next() {
-		var (
-			login  string
-			id     int64
-			amount float64
-			date   time.Time
-		)
-		err := rows.Scan(&login, &id, &amount, &date)
-		if err != nil {
-			return nil, err
-		}
-		transactions = append(transactions, models.Transaction{ID: id, Amount: amount, User: login, Date: date})
-	}
-	return transactions, nil
 }
